add api of user login

+++ fim_auth/auth_api/internal/config/config.go
@@ -11,6 +11,6 @@ type Config struct {
 	}
 	Auth struct {
 		AccessSecret string
-		AccessExpire int64
+		AccessExpire int
 	}
 }
